+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project **may some day** to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.2] - 2019-12-03
+### Added
+- Shell#attempt will execute a script and ignore errors
+
+### Changed
+- Shell#check_point now executes scripts
+- Shell#transact no longer allows chaining with other commands
+
 ## [0.0.1] - 2019-12-02
 ### Added
 - Initial version of ShellB